requirements/local.txt
-r base.txt

django_debug_toolbar==3.8.1